Html

Documentation

module Html.Type

module Html.Convert

renderString :: Document a => a -> String

Render a html document to a String.

renderText :: Document a => a -> Text

Render a html document to a lazy Text.

renderByteString :: Document a => a -> ByteString

Render a html document to a lazy ByteString.

renderBuilder :: Document a => a -> Builder

Render a html document to a Builder.

compactHTML :: Compactable a => a -> CompactHTML (Variables a)

Compact a html document.

renderCompactString :: Retrievable a => CompactHTML a -> Retrieve String a

Render a compacted html document to a String.

renderCompactText :: Retrievable a => CompactHTML a -> Retrieve Text a

Render a compacted html document to a lazy Text.

renderCompactByteString :: Retrievable a => CompactHTML a -> Retrieve ByteString a

Render a compacted html document to a lazy ByteString.

renderCompactBuilder :: Retrievable a => CompactHTML a -> Retrieve Builder a

Render a compacted html document to a Builder.

data Put (n :: Symbol)

Data for putting variables into a rendered compacted html document.

Constructors

forall a.Convert a => Put a

data V (n :: Symbol)

Data for declaring variables in a html document which will be compacted.

Constructors

V

Orphan instances

Document (a := b) => Show (a := b)
Show instances to faciliate ghci development.

Instance details

Methods

showsPrec :: Int -> (a := b) -> ShowS

show :: (a := b) -> String

showList :: [a := b] -> ShowS

Document (a :@ b) => Show (a :@ b)

Instance details

Methods

showsPrec :: Int -> (a :@ b) -> ShowS

show :: (a :@ b) -> String

showList :: [a :@ b] -> ShowS

Document (a :> b) => Show (a :> b)

Instance details

Methods

showsPrec :: Int -> (a :> b) -> ShowS

show :: (a :> b) -> String

showList :: [a :> b] -> ShowS

Document (a # b) => Show (a # b)

Instance details

Methods

showsPrec :: Int -> (a # b) -> ShowS

show :: (a # b) -> String

showList :: [a # b] -> ShowS

Document (Attribute a global boolean) => Show (Attribute a global boolean)

Instance details

Methods

showsPrec :: Int -> Attribute a global boolean -> ShowS

show :: Attribute a global boolean -> String

showList :: [Attribute a global boolean] -> ShowS

Document (Element name categories contentModel contentAttributes) => Show (Element name categories contentModel contentAttributes)

Instance details

Methods

showsPrec :: Int -> Element name categories contentModel contentAttributes -> ShowS

show :: Element name categories contentModel contentAttributes -> String

showList :: [Element name categories contentModel contentAttributes] -> ShowS